requirements.txt
pandas
yfinance
requests
finvizfinance
paramiko
lxmlschedule
